@@ -1,6 +1,8 @@
-export * from './src/dispatcher';
-export * from './src/ng2';
-export * from './src/reducer';
-export * from './src/state';
-export * from './src/store';
-export * from './src/utils';
+import * as __private_export__ from './src/private_export';
+
+export { Action, ActionReducer, ActionReducerMap, ActionReducerFactory } from './src/models';
+export { StoreModule } from './src/store_module';
+export { Store } from './src/store';
+export { combineReducers, compose } from './src/utils';
+export { __private_export__ };
+
