@@ -1,3 +1,9 @@
+export interface TodoItem {
+  id: number;
+  completed: boolean;
+  text: string;
+}
+
 export const ADD_TODO = 'ADD_TODO';
 export const COMPLETE_TODO = 'COMPLETE_TODO';
 export const SET_VISIBILITY_FILTER = 'SET_VISIBILITY_FILTER';
@@ -20,7 +26,7 @@ export function visibilityFilter(state = VisibilityFilters.SHOW_ALL, {type, payl
   }
 };
 
-export function todos(state = [], {type, payload}) {
+export function todos(state: TodoItem[] = [], {type, payload}): TodoItem[] {
   switch (type) {
     case ADD_TODO:
       return [
@@ -32,11 +38,11 @@ export function todos(state = [], {type, payload}) {
         }
       ];
     case COMPLETE_ALL_TODOS:
-      return state.map(todo => Object.assign({}, todo, {completed: true}));
+      return state.map(todo => ({ ...todo, completed: true }));
     case COMPLETE_TODO:
-      return state.map(todo => {
-        return todo.id === payload.id ? Object.assign({}, todo, {completed: true}) : todo;
-      });
+      return state.map(todo =>
+        todo.id === payload.id ? { ...todo, completed: true } : todo
+      );
     default:
       return state;
   }
