@@ -1,16 +1,35 @@
 import { NgModule, Component } from '@angular/core';
 import { platformDynamicServer } from '@angular/platform-server';
 import { BrowserModule } from '@angular/platform-browser';
-import { Store, StoreModule } from '../../';
+import { Store, StoreModule, combineReducers } from '../../';
 import { counterReducer, INCREMENT, DECREMENT } from '../fixtures/counter';
+import { todos } from '../fixtures/todos';
 import { Observable } from 'rxjs/Observable';
+
+@Component({
+  selector: 'ngc-spec-child-component',
+  template: `
+  
+  `
+})
+export class NgcSpecChildComponent { }
+
+@NgModule({
+  imports: [
+    StoreModule.forFeature('feature', { todos: todos })
+  ],
+  declarations: [
+    NgcSpecChildComponent,
+  ],
+  exports: [
+    NgcSpecChildComponent,
+  ]
+})
+export class FeatureModule { }
 
 export interface AppState {
   count: number;
 }
-
-export const storeConfig = {count: counterReducer};
-export const initialState = { count : 0 };
 
 @Component({
   selector: 'ngc-spec-component',
@@ -18,6 +37,8 @@ export const initialState = { count : 0 };
     <button (click)="increment()"> + </button>
     <span>  Count : {{ count | async }}  </span>
     <button (click)="decrement()"> + </button>
+
+    <ngc-spec-child-component></ngc-spec-child-component>
   `
 })
 export class NgcSpecComponent {
@@ -36,7 +57,11 @@ export class NgcSpecComponent {
 @NgModule({
   imports: [
     BrowserModule,
-    StoreModule.provideStore(storeConfig, initialState)
+    StoreModule.forRoot({ count: counterReducer }, {
+      initialState: { count : 0 },
+      reducerFactory: combineReducers
+    }),
+    FeatureModule
   ],
   declarations: [NgcSpecComponent],
   bootstrap: [NgcSpecComponent]
